requirements.txt
aiogram>=3.0.0
python-dotenv>=0.21.0
openai>=1.0.0
httpx>=0.24.0
fastapi>=0.122.0
uvicorn>=0.38.0